nonexistent socket file isn't fatal
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@1490 406a0c4d-033a-0410-8de8-e80135713968

+++ src/searchd.cpp
@@ -948,7 +948,10 @@ int sphCreateUnixSocket ( const char * sPath )
 		sphFatal ( "failed to create UNIX socket: %s", sphSockError() );
 
 	if ( unlink ( sPath ) == -1 )
-		sphFatal ( "unlink() on UNIX socket file failed: %s", sphSockError() );
+	{
+		if ( errno != ENOENT )
+			sphFatal ( "unlink() on UNIX socket file failed: %s", sphSockError() );
+	}
 	
 	if ( bind ( iSock, (struct sockaddr *)&uaddr, sizeof(uaddr) ) != 0 )
 		sphFatal ( "bind() on UNIX socket failed: %s", sphSockError() );
